separate file mirror and manager from main
