+++ main.py
@@ -28,13 +28,13 @@ from storj import model
 #from client_configuration_ui import Ui_ClientConfiguration
 #from create_bucket_ui import Ui_BucketCreate
 from file_crypto_tools import FileCrypto  # file ancryption and decryption lib
-from file_manager_ui import Ui_FileManager
-from file_mirrors_list_ui import Ui_FileMirrorsList
+#from file_manager_ui import Ui_FileManager
+#from file_mirrors_list_ui import Ui_FileMirrorsList
 #from initial_window_ui import Ui_InitialWindow
 #from main_menu_ui import Ui_MainMenu
 #from node_details_ui import Ui_NodeDetails
 #from single_file_downloader_ui import Ui_SingleFileDownload
-from single_file_upload_ui import Ui_SingleFileUpload
+#from single_file_upload_ui import Ui_SingleFileUpload
 #from storj_login_ui import Ui_Login
 #from storj_register_ui import Ui_Register
 from sharder import ShardingTools
@@ -124,170 +124,6 @@ class MyTableModel(QtCore.QAbstractTableModel):
 
 
 
-# Mirrors section
-class FileMirrorsListUI(QtGui.QMainWindow):
-    def __init__(self, parent=None, bucketid=None, fileid=None):
-        QtGui.QWidget.__init__(self, parent)
-        self.file_mirrors_list_ui = Ui_FileMirrorsList()
-        self.file_mirrors_list_ui.setupUi(self)
-        # model = self.file_mirrors_list_ui.established_mirrors_tree.model()
-
-
-        self.file_mirrors_list_ui.mirror_details_bt.clicked.connect(
-            lambda: self.open_mirror_details_window("established"))
-        self.file_mirrors_list_ui.mirror_details_bt_2.clicked.connect(
-            lambda: self.open_mirror_details_window("available"))
-        self.file_mirrors_list_ui.quit_bt.clicked.connect(self.close)
-
-        # self.connect(self.file_mirrors_list_ui.established_mirrors_tree, QtCore.SIGNAL("selectionChanged(QItemSelection, QItemSelection)"), self.open_mirror_details_window)
-
-        # self.connect(self.file_mirrors_list_ui.established_mirrors_tree, QtCore.SIGNAL('selectionChanged()'), self.open_mirror_details_window)
-
-
-        # QtCore.QObject.connect(self.file_mirrors_list_ui.established_mirrors_tree.selectionModel(), QtCore.SIGNAL('selectionChanged(QItemSelection, QItemSelection)'),
-        # self.open_mirror_details_window)
-
-
-        # self.file_mirrors_list_ui.established_mirrors_tree.
-
-        self.bucketid = bucketid
-        self.fileid = fileid
-
-        self.file_mirrors_list_ui.file_id_label.setText(html_format_begin + str(self.fileid) + html_format_end)
-
-        print self.fileid
-        self.storj_engine = StorjEngine()  # init StorjEngine
-        self.createNewMirrorListInitializationThread()
-
-    def open_mirror_details_window(self, mirror_state):
-        # self.established_mirrors_tree_view = self.file_mirrors_list_ui.established_mirrors_tree
-
-
-        # daat = self.file_mirrors_list_ui.established_mirrors_tree.selectedIndexes()
-        # model = self.file_mirrors_list_ui.established_mirrors_tree.model()
-        # data = []
-
-        # initialize variables
-        item = ""
-        index = ""
-        try:
-            if mirror_state == "established":
-                index = self.file_mirrors_list_ui.established_mirrors_tree.selectedIndexes()[3]
-                item = self.file_mirrors_list_ui.established_mirrors_tree.selectedIndexes()[3]
-            elif mirror_state == "available":
-                index = self.file_mirrors_list_ui.available_mirrors_tree.selectedIndexes()[3]
-                item = self.file_mirrors_list_ui.available_mirrors_tree.selectedIndexes()[3]
-
-            nodeid_to_send = item.model().itemFromIndex(index).text()
-
-            if nodeid_to_send != "":
-                self.node_details_window = NodeDetailsUI(self, nodeid_to_send)
-                self.node_details_window.show()
-            else:
-                QMessageBox.about(self, "Warning", "Please select farmer node from list")
-                print "Unhandled error"
-
-        except:
-            QMessageBox.about(self, "Warning", "Please select farmer node from list")
-            print "Unhandled error"
-
-    def createNewMirrorListInitializationThread(self):
-        mirror_list_initialization_thread = threading.Thread(target=self.initialize_mirrors_tree, args=())
-        mirror_list_initialization_thread.start()
-
-    def initialize_mirrors_tree(self):
-        # create model
-        # model = QtGui.QFileSystemModel()
-        # model.setRootPath(QtCore.QDir.currentPath())
-
-        self.file_mirrors_list_ui.loading_label_mirrors_established.setStyleSheet('color: red')  # set loading color
-        self.file_mirrors_list_ui.loading_label_mirrors_available.setStyleSheet('color: red')  # set loading color
-
-        self.mirror_tree_view_header = ['Shard Hash / Address', 'User agent', 'Last seed', 'Node ID']
-
-        ######################### set the model for established mirrors ##################################
-        self.established_mirrors_model = QStandardItemModel()
-        self.established_mirrors_model.setHorizontalHeaderLabels(self.mirror_tree_view_header)
-
-        self.established_mirrors_tree_view = self.file_mirrors_list_ui.established_mirrors_tree
-        self.established_mirrors_tree_view.setSelectionBehavior(QAbstractItemView.SelectRows)
-
-        self.established_mirrors_tree_view.setModel(self.established_mirrors_model)
-        self.established_mirrors_tree_view.setUniformRowHeights(True)
-
-        self.file_mirrors_list_ui.available_mirrors_tree.setModel(self.established_mirrors_model)
-
-        divider = 0
-        group = 1
-        self.established_mirrors_count_for_file = 0
-        recent_shard_hash = "";
-        parent1 = QStandardItem('')
-        for file_mirror in self.storj_engine.storj_client.file_mirrors(str(self.bucketid), str(self.fileid)):
-            for mirror in file_mirror.established:
-                self.established_mirrors_count_for_file += 1
-                print file_mirror.established
-                if mirror["shardHash"] != recent_shard_hash:
-                    parent1 = QStandardItem('Shard with hash {}'.format(mirror["shardHash"]))
-                    divider = divider + 1
-                    self.established_mirrors_model.appendRow(parent1)
-
-                child1 = QStandardItem(str(mirror["contact"]["address"] + ":" + str(mirror["contact"]["port"])))
-                child2 = QStandardItem(str(mirror["contact"]["userAgent"]))
-                child3 = QStandardItem(str(mirror["contact"]["lastSeen"]))
-                child4 = QStandardItem(str(mirror["contact"]["nodeID"]))
-                parent1.appendRow([child1, child2, child3, child4])
-
-                # span container columns
-                # self.established_mirrors_tree_view.setFirstColumnSpanned(1, self.established_mirrors_tree_view.rootIndex(), True)
-
-                recent_shard_hash = mirror["shardHash"]
-
-        self.file_mirrors_list_ui.loading_label_mirrors_established.setText("")
-
-        # dbQueryModel.itemData(treeView.selectedIndexes()[0])
-
-        ################################### set the model for available mirrors #########################################
-        self.available_mirrors_model = QStandardItemModel()
-        self.available_mirrors_model.setHorizontalHeaderLabels(self.mirror_tree_view_header)
-
-        self.available_mirrors_tree_view = self.file_mirrors_list_ui.available_mirrors_tree
-        self.available_mirrors_tree_view.setSelectionBehavior(QAbstractItemView.SelectRows)
-
-        self.available_mirrors_tree_view.setModel(self.available_mirrors_model)
-        self.available_mirrors_tree_view.setUniformRowHeights(True)
-
-        self.file_mirrors_list_ui.available_mirrors_tree.setModel(self.available_mirrors_model)
-
-        divider = 0
-        self.available_mirrors_count_for_file = 0
-        recent_shard_hash_2 = "";
-        parent2 = QStandardItem('')
-        for file_mirror in self.storj_engine.storj_client.file_mirrors(str(self.bucketid), str(self.fileid)):
-            for mirror_2 in file_mirror.available:
-                self.available_mirrors_count_for_file += 1
-                if mirror_2["shardHash"] != recent_shard_hash_2:
-                    parent2 = QStandardItem('Shard with hash {}'.format(mirror_2["shardHash"]))
-                    divider = divider + 1
-                    self.available_mirrors_model.appendRow(parent2)
-
-                child1 = QStandardItem(str(mirror_2["contact"]["address"] + ":" + str(mirror_2["contact"]["port"])))
-                child2 = QStandardItem(str(mirror_2["contact"]["userAgent"]))
-                child3 = QStandardItem(str(mirror_2["contact"]["lastSeen"]))
-                child4 = QStandardItem(str(mirror_2["contact"]["nodeID"]))
-                parent2.appendRow([child1, child2, child3, child4])
-
-                # span container columns
-                # self.established_mirrors_tree_view.setFirstColumnSpanned(1, self.established_mirrors_tree_view.rootIndex(), True)
-
-                recent_shard_hash_2 = mirror_2["shardHash"]
-        self.file_mirrors_list_ui.loading_label_mirrors_available.setText("")
-
-        self.file_mirrors_list_ui.established_mirrors_count.setText(
-            html_format_begin + str(self.established_mirrors_count_for_file) + html_format_end)
-        self.file_mirrors_list_ui.available_mirrors_count.setText(
-            html_format_begin + str(self.available_mirrors_count_for_file) + html_format_end)
-        print QtCore.QDir.currentPath()
-
 
 
 ######################################################################################################################################
@@ -350,153 +186,6 @@ class BucketCreateUI(QtGui.QMainWindow):
 
 ######################################################################################################################################
 ####################### FILE MANAGER UI ##################################
-
-# Files section
-class FileManagerUI(QtGui.QMainWindow):
-    def __init__(self, parent=None, bucketid=None):
-        QtGui.QWidget.__init__(self, parent)
-        self.file_manager_ui = Ui_FileManager()
-        self.file_manager_ui.setupUi(self)
-
-        QtCore.QObject.connect(self.file_manager_ui.bucket_select_combo_box,
-                               QtCore.SIGNAL("currentIndexChanged(const QString&)"),
-                               self.createNewFileListUpdateThread)  # connect ComboBox change listener
-        QtCore.QObject.connect(self.file_manager_ui.file_mirrors_bt, QtCore.SIGNAL("clicked()"),
-                               self.open_mirrors_list_window)  # create bucket action
-        QtCore.QObject.connect(self.file_manager_ui.quit_bt, QtCore.SIGNAL("clicked()"),
-                               self.close)  # create bucket action
-        QtCore.QObject.connect(self.file_manager_ui.file_download_bt, QtCore.SIGNAL("clicked()"),
-                               self.open_single_file_download_window)  # create bucket action
-        QtCore.QObject.connect(self.file_manager_ui.file_delete_bt, QtCore.SIGNAL("clicked()"),
-                               self.delete_selected_file)  # delete selected file
-
-        self.storj_engine = StorjEngine()  # init StorjEngine
-        self.createNewBucketResolveThread()
-
-    def delete_selected_file(self):
-
-
-        self.current_bucket_index = self.file_manager_ui.bucket_select_combo_box.currentIndex()
-        self.current_selected_bucket_id = self.bucket_id_list[self.current_bucket_index]
-
-        tablemodel = self.file_manager_ui.files_list_tableview.model()
-        rows = sorted(set(index.row() for index in
-                          self.file_manager_ui.files_list_tableview.selectedIndexes()))
-
-        selected = False
-        for row in rows:
-            selected = True
-            index = tablemodel.index(row, 3)  # get file ID index
-            index_filename = tablemodel.index(row, 0)  # get file name index
-
-            # We suppose data are strings
-            selected_file_id = str(tablemodel.data(index).toString())
-            selected_file_name = str(tablemodel.data(index_filename).toString())
-            msgBox = QtGui.QMessageBox(QtGui.QMessageBox.Question, "Question",
-                                       "Are you sure you want to delete this file? File name: " + selected_file_name, QtGui.QMessageBox.Yes)
-            result = msgBox.exec_()
-            print result
-            if result != QtGui.QMessageBox.Rejected:
-                try:
-                    self.storj_engine.storj_client.file_remove(str(self.current_selected_bucket_id), str(selected_file_id))
-                    self.createNewFileListUpdateThread() # update files list
-                    QMessageBox.about(self, "Success", 'File "' + str(selected_file_name) + '" was deleted successfully')
-                except storj.exception.StorjBridgeApiError, e:
-                    QMessageBox.about(self, "Error", "Bridge exception occured while trying to delete file: " + str(e))
-                except Exception, e:
-                    QMessageBox.about(self, "Error", "Unhandled exception occured while trying to delete file: " + str(e))
-
-        if not selected:
-            QMessageBox.about(self, "Information", "Please select file which you want to delete")
-
-        return True
-
-    def open_mirrors_list_window(self):
-        self.current_bucket_index = self.file_manager_ui.bucket_select_combo_box.currentIndex()
-        self.current_selected_bucket_id = self.bucket_id_list[self.current_bucket_index]
-
-        tablemodel = self.file_manager_ui.files_list_tableview.model()
-        rows = sorted(set(index.row() for index in
-                          self.file_manager_ui.files_list_tableview.selectedIndexes()))
-        i = 0
-        for row in rows:
-            print('Row %d is selected' % row)
-            index = tablemodel.index(row, 3)  # get file ID
-            # We suppose data are strings
-            selected_file_id = str(tablemodel.data(index).toString())
-            self.file_mirrors_list_window = FileMirrorsListUI(self, str(self.current_selected_bucket_id),
-                                                              selected_file_id)
-            self.file_mirrors_list_window.show()
-            i += 1
-
-        if i == 0:
-            QMessageBox.about(self, "Warning!", "Please select file from file list!")
-
-        print 1
-
-    def createNewFileListUpdateThread(self):
-        download_thread = threading.Thread(target=self.update_files_list, args=())
-        download_thread.start()
-
-    def update_files_list(self):
-
-        self.tools = Tools()
-
-        model = QStandardItemModel(1, 1)  # initialize model for inserting to table
-
-        model.setHorizontalHeaderLabels(['File name', 'File size', 'Mimetype', 'File ID'])
-
-        self.current_bucket_index = self.file_manager_ui.bucket_select_combo_box.currentIndex()
-        self.current_selected_bucket_id = self.bucket_id_list[self.current_bucket_index]
-
-        i = 0
-
-        for self.file_details in self.storj_engine.storj_client.bucket_files(str(self.current_selected_bucket_id)):
-            item = QStandardItem(str(self.file_details["filename"]))
-            model.setItem(i, 0, item)  # row, column, item (QStandardItem)
-
-            file_size_str = self.tools.human_size(int(self.file_details["size"])) # get human readable file size
-
-            item = QStandardItem(str(file_size_str))
-            model.setItem(i, 1, item)  # row, column, item (QStandardItem)
-
-            item = QStandardItem(str(self.file_details["mimetype"]))
-            model.setItem(i, 2, item)  # row, column, item (QStandardItem)
-
-            item = QStandardItem(str(self.file_details["id"]))
-            model.setItem(i, 3, item)  # row, column, item (QStandardItem)
-
-            i = i + 1
-
-            print self.file_details
-
-        self.file_manager_ui.files_list_tableview.clearFocus()
-        self.file_manager_ui.files_list_tableview.setModel(model)
-        self.file_manager_ui.files_list_tableview.horizontalHeader().setResizeMode(QtGui.QHeaderView.Stretch)
-
-    def createNewBucketResolveThread(self):
-        download_thread = threading.Thread(target=self.initialize_bucket_select_combobox, args=())
-        download_thread.start()
-
-    def initialize_bucket_select_combobox(self):
-        self.buckets_list = []
-        self.bucket_id_list = []
-        self.storj_engine = StorjEngine()  # init StorjEngine
-        i = 0
-        try:
-            for bucket in self.storj_engine.storj_client.bucket_list():
-                self.buckets_list.append(str(bucket.name))  # append buckets to list
-                self.bucket_id_list.append(str(bucket.id))  # append buckets to list
-                i = i + 1
-        except storj.exception.StorjBridgeApiError, e:
-            QMessageBox.about(self, "Unhandled bucket resolving exception", "Exception: " + str(e))
-
-        self.file_manager_ui.bucket_select_combo_box.addItems(self.buckets_list)
-
-    def open_single_file_download_window(self):
-        self.single_file_download_window = SingleFileDownloadUI(self)
-        self.single_file_download_window.show()
-
